+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Add check to prevent re-initialising if semver file already exists
 - Add configuration options for semver file
 - Change rake task namespace from `semantic_release` to `release` and make configurable